requirements.txt
Flask
python-dotenv
groq
openai
